登陆流程调整v1

+++ server/Base/Timer.h
@@ -11,11 +11,11 @@ public:
 	CTimer();
 	~CTimer();
 	
-	static uint32 GetTime() { return m_Time; }
+	static int64 GetTime() { return m_Time; }
 	static uint64 GetTime64() { return m_Time64; }
 	static void UpdateTime();
 private:
 
-	static uint32 m_Time;
+	static int64 m_Time;
 	static uint64 m_Time64;
 };
